# Do not use defer for unlocking in Scratch