@bep 2016-03-24
Also see: https://github.com/golang/go/issues/14939

For future reference: `defer` is a brilliant construct for closing files etc.; but if you use it in high-frequent places (millions ... billions of invocations?) it may be worth to do it explicit. 109 nanoseconds is a short time span, but it adds up.